fix: always apply vite ssr source maps
